moves conduit to protocol package
